Only generate relative links to files that exist
